Move logic of expanding cells from TableViewDataSource to ExpandableCellsHandler
